refactored ConditionsSections into page folder
