+++ dfm-sideline-sidekick-app/App.tsx
@@ -7,7 +7,7 @@ import React from "react";
 import { StyleSheet } from "react-native";
 
 import AppInitializer from "./AppInitializer";
-import ConditionsSection from "./ConditionsSection";
+import ConditionsSection from "./pages/ConditionsSection";
 import { DataProvider } from "./DataContext";
 import { BottomNavBar, NavItem } from "./components/bar";
 import BookmarkPage from "./pages/BookmarkPage";
