+++ dfm-sideline-sidekick-app/components/StringRenderer.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Text, View } from "react-native";
 
-import styles from "../ConditionSectionStyles";
+import styles from "../pages/ConditionSectionStyles";
 
 type StringValue = string | string[] | { [key: string]: StringValue } | undefined;
