+++ dfm-sideline-sidekick-app/pages/ConditionSectionStyles.tsx
